backend: Move telemetry files to their own folder
