+++ packages/backend/src/ats/lever.ts
@@ -1,9 +1,9 @@
 import { standardizeUntrustedHtml } from "dry-utils-text";
 import { config } from "../config.ts";
 import type { Company, CompanyKey, Job, JobKey } from "../models/models.ts";
+import { logError } from "../telemetry/telemetry.ts";
 import type { Context } from "../types/types.ts";
 import { AppError } from "../utils/AppError.ts";
-import { logError } from "../utils/telemetry.ts";
 import { ATSBase } from "./atsBase.ts";
 import { normTitle } from "./normalization.ts";
