+++ packages/backend/src/controllers/company.ts
@@ -5,9 +5,9 @@ import { db } from "../db/db.ts";
 import type { RefreshJobsOptions } from "../models/clientModels.ts";
 import type { CompanyKey, CompanyKeys, FullJobKey } from "../models/models.ts";
 import { refreshJobsForCompany } from "../sync/jobs.ts";
+import { logProperty } from "../telemetry/telemetry.ts";
 import { AppError } from "../utils/AppError.ts";
 import { AsyncQueue } from "../utils/asyncQueue.ts";
-import { logProperty } from "../utils/telemetry.ts";
 import { refreshMetadata } from "./metadata.ts";
 
 const companyInfoQueue = new AsyncQueue(
