+++ packages/backend/src/sync/jobs.ts
@@ -3,10 +3,10 @@ import { ats } from "../ats/ats.ts";
 import { refreshMetadata } from "../controllers/metadata.ts";
 import { db } from "../db/db.ts";
 import type { CompanyKey, Job } from "../models/models.ts";
+import { logProperty } from "../telemetry/telemetry.ts";
 import type { Context } from "../types/types.ts";
 import { AsyncQueue } from "../utils/asyncQueue.ts";
 import { JOB_EXPIRY_MS } from "../utils/constants.ts";
-import { logProperty } from "../utils/telemetry.ts";
 
 const jobInfoQueue = new AsyncQueue("RefreshJobInfo", refreshJobInfo, {
   onComplete: refreshMetadata,
