+++ packages/backend/src/telemetry/telemetryWorkaround.cjs
