+++ packages/backend/test/setup.ts
@@ -11,7 +11,7 @@ process.env.NODE_ENV ??= "test";
 process.env.OPENAI_API_KEY ??= "test-openai-key-123456";
 
 const telemetryWorkaround = (
-  await import("../src/utils/telemetryWorkaround.cjs")
+  await import("../src/telemetry/telemetryWorkaround.cjs")
 ).default;
 const telemetryClient = {
   addTelemetryProcessor: () => {},
